small fix (if map is smaller than viewport + infinite map)
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@4600 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ src/client/GfxPrimitives.cpp
@@ -511,7 +511,7 @@ void DrawImageTiled(SDL_Surface *bmpDest, SDL_Surface *bmpSrc, int sx, int sy, i
 {
 	SDL_Rect newClip = {dx,dy,dw,dh};
 	ScopedSurfaceClip clip(bmpDest,newClip);	
-	MOD(sx, bmpSrc->w); MOD(sy, bmpSrc->h);
+	MOD(sx, (int)bmpSrc->w); MOD(sy, (int)bmpSrc->h);
 	
 	// Place the tiles
 	for (int y = dy; y < dy + dh; y += sh)
